chore(node): rename event_bus to event
